refactor: move import tests with the  rest of tests
