+++ crates/tx3-lang/src/parsing.rs
@@ -2833,4 +2833,6 @@ mod tests {
     test_parsing!(list_concat);
 
     test_parsing!(buidler_fest_2026);
+
+    test_parsing!(imported_datum);
 }
