--- crates/tx3-lang/tests/fixtures/import_test.tx3
@@ -1,6 +1,0 @@
-import "../../../cip-57/examples/plutus.json" as types;
-
-party Sender;
-party Receiver;
-
-tx dummy() {}
